EXTENDRAT-37: Refactor the code and improve the data model to facilitate the creation of rating types

+++ src/java/fr/paris/lutece/plugins/extend/modules/rating/service/validator/IRatingValidationService.java
@@ -54,5 +54,5 @@ public interface IRatingValidationService
      *         resource, or null if he is allowed
      */
     String validateRating( HttpServletRequest request, LuteceUser user, String strIdResource,
-        String strResourceTypeKey, double dVoteValue );
+        String strResourceTypeKey, float dVoteValue );
 }
